requirements.txt
aiogram~=2.25.1
python-decouple~=3.8
psycopg2-binary